[API-1863] Disallow modifying generic records after they are built (#23948)
This commit changes only the compact related files because in portable
generic records there is a class definition and users cannot write any
other field defined in that definition. Plus, there is a check which
fields are written, it is impossible to write a field twice. These mean
a user won't be able to modify a Portable generic record after building
it.

This commit also fixes a bug. DeserializedGenericRecord's
newBuilderWithClone method was not copying the 'objects' field, leading
to changing the original generic record if there is a change while
cloning it.

Fixes API-1863
