+++ hazelcast/src/main/java/com/hazelcast/nio/serialization/genericrecord/GenericRecordBuilder.java
@@ -105,6 +105,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setBoolean(@Nonnull String fieldName, boolean value);
@@ -127,6 +130,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setInt8(@Nonnull String fieldName, byte value);
@@ -152,6 +158,9 @@ public interface GenericRecordBuilder {
      *                                         set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setChar(@Nonnull String fieldName, char value);
@@ -174,6 +183,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setInt16(@Nonnull String fieldName, short value);
@@ -196,6 +208,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setInt32(@Nonnull String fieldName, int value);
@@ -218,6 +233,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setInt64(@Nonnull String fieldName, long value);
@@ -240,6 +258,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setFloat32(@Nonnull String fieldName, float value);
@@ -262,6 +283,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setFloat64(@Nonnull String fieldName, double value);
@@ -285,6 +309,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableBoolean(@Nonnull String fieldName, @Nullable Boolean value);
@@ -308,6 +335,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableInt8(@Nonnull String fieldName, @Nullable Byte value);
@@ -332,6 +362,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableInt16(@Nonnull String fieldName, @Nullable Short value);
@@ -355,6 +388,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableInt32(@Nonnull String fieldName, @Nullable Integer value);
@@ -378,6 +414,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableInt64(@Nonnull String fieldName, @Nullable Long value);
@@ -401,6 +440,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableFloat32(@Nonnull String fieldName, @Nullable Float value);
@@ -424,6 +466,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setNullableFloat64(@Nonnull String fieldName, @Nullable Double value);
@@ -446,6 +491,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setString(@Nonnull String fieldName, @Nullable String value);
@@ -473,6 +521,9 @@ public interface GenericRecordBuilder {
      *                                         the same as the generic record that is
      *                                         being built. e.g using portable generic
      *                                         record in a compact generic record builder.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setGenericRecord(@Nonnull String fieldName, @Nullable GenericRecord value);
@@ -498,6 +549,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setDecimal(@Nonnull String fieldName, @Nullable BigDecimal value);
@@ -522,6 +576,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setTime(@Nonnull String fieldName, @Nullable LocalTime value);
@@ -547,6 +604,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setDate(@Nonnull String fieldName, @Nullable LocalDate value);
@@ -572,6 +632,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setTimestamp(@Nonnull String fieldName, @Nullable LocalDateTime value);
@@ -598,6 +661,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setTimestampWithTimezone(@Nonnull String fieldName, @Nullable OffsetDateTime value);
@@ -620,6 +686,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfBoolean(@Nonnull String fieldName, @Nullable boolean[] value);
@@ -642,6 +711,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfInt8(@Nonnull String fieldName, @Nullable byte[] value);
@@ -667,6 +739,9 @@ public interface GenericRecordBuilder {
      *                                         set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfChar(@Nonnull String fieldName, @Nullable char[] value);
@@ -689,6 +764,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfInt16(@Nonnull String fieldName, @Nullable short[] value);
@@ -711,6 +789,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfInt32(@Nonnull String fieldName, @Nullable int[] value);
@@ -733,6 +814,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfInt64(@Nonnull String fieldName, @Nullable long[] value);
@@ -755,6 +839,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfFloat32(@Nonnull String fieldName, @Nullable float[] value);
@@ -777,6 +864,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfFloat64(@Nonnull String fieldName, @Nullable double[] value);
@@ -800,6 +890,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableBoolean(@Nonnull String fieldName, @Nullable Boolean[] value);
@@ -823,6 +916,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableInt8(@Nonnull String fieldName, @Nullable Byte[] value);
@@ -846,6 +942,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableInt16(@Nonnull String fieldName, @Nullable Short[] value);
@@ -869,6 +968,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableInt32(@Nonnull String fieldName, @Nullable Integer[] value);
@@ -892,6 +994,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableInt64(@Nonnull String fieldName, @Nullable Long[] value);
@@ -915,6 +1020,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableFloat32(@Nonnull String fieldName, @Nullable Float[] value);
@@ -938,6 +1046,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfNullableFloat64(@Nonnull String fieldName, @Nullable Double[] value);
@@ -960,6 +1071,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfString(@Nonnull String fieldName, @Nullable String[] value);
@@ -982,6 +1096,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      * @see #setDecimal(String, BigDecimal)
      */
     @Nonnull
@@ -1005,6 +1122,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      * @see #setTime(String, LocalTime)
      */
     @Nonnull
@@ -1028,6 +1148,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      * @see #setDate(String, LocalDate)
      */
     @Nonnull
@@ -1051,6 +1174,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      * @see #setTimestamp(String, LocalDateTime)
      */
     @Nonnull
@@ -1074,6 +1200,9 @@ public interface GenericRecordBuilder {
      *                                         trying to be set without using
      *                                         {@link
      *                                         GenericRecord#newBuilderWithClone()}.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      * @see #setTimestampWithTimezone(String, OffsetDateTime)
      */
     @Nonnull
@@ -1108,6 +1237,9 @@ public interface GenericRecordBuilder {
      *                                         the same as the generic record that is
      *                                         being built. e.g using portable generic
      *                                         record in a compact generic record builder.
+     * @throws UnsupportedOperationException   if the setter is called after a
+     *                                         GenericRecord is built by
+     *                                         {@link GenericRecordBuilder#build}.
      */
     @Nonnull
     GenericRecordBuilder setArrayOfGenericRecord(@Nonnull String fieldName, @Nullable GenericRecord[] value);
